# Eliminate allocation in `StreamAsIStream.Read`