--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/StreamAsIStream.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/StreamAsIStream.cs
@@ -10,6 +10,7 @@ using System.Security;
 using System;
 using MS.Internal;
 using MS.Win32;
+using System.Buffers;
 using System.Reflection;
 using System.Collections;
 using System.Diagnostics;
@@ -28,7 +29,7 @@ namespace System.Windows.Media
     internal struct StreamDescriptor
     {
         internal delegate void Dispose(ref StreamDescriptor pSD);
-        internal delegate int Read(ref StreamDescriptor pSD, [MarshalAs(UnmanagedType.LPArray, SizeParamIndex = 2), Out]byte[] buffer, uint cb, out uint cbRead);
+        internal delegate int Read(ref StreamDescriptor pSD, IntPtr buffer, uint cb, out uint cbRead);
 
         internal unsafe delegate int Seek(ref StreamDescriptor pSD, long offset, uint origin, long* plibNewPostion);
         internal delegate int Stat(ref StreamDescriptor pSD, out System.Runtime.InteropServices.ComTypes.STATSTG statstg, uint grfStatFlag);
@@ -237,7 +238,7 @@ namespace System.Windows.Media
 
                     uint read = 0;
 
-                    hr = Read(buffer, toRead, out read);
+                    hr = Read(buffer.AsSpan(0, (int) toRead), out read);
 
                     if (read == 0)
                     {
@@ -290,7 +291,7 @@ namespace System.Windows.Media
             return NativeMethods.E_NOTIMPL;
         }
 
-        public int Read(byte[] buffer, uint cb, out uint cbRead)
+        public int Read(Span<byte> buffer, out uint cbRead)
         {
             cbRead = 0;
 
@@ -301,7 +302,7 @@ namespace System.Windows.Media
                 Verify();
                 ActualizeVirtualPosition();
 
-                cbRead = (uint) dataStream.Read(buffer, 0, (int) cb);
+                cbRead = (uint) dataStream.Read(buffer);
             }
             catch (Exception e)
             {
@@ -597,9 +598,10 @@ namespace System.Windows.Media
             return (StreamAsIStream.FromSD(ref pSD)).LockRegion(libOffset, cb, dwLockType);
         }
 
-        internal static int Read(ref StreamDescriptor pSD, byte[] buffer, uint cb, out uint cbRead)
+        internal static unsafe int Read(ref StreamDescriptor pSD, IntPtr buffer, uint cb, out uint cbRead)
         {
-            return (StreamAsIStream.FromSD(ref pSD)).Read(buffer, cb, out cbRead);
+            var span = new Span<byte>(buffer.ToPointer(), (int) cb);
+            return (StreamAsIStream.FromSD(ref pSD)).Read(span, out cbRead);
         }
 
         internal static int Revert(ref StreamDescriptor pSD)