## Description

The native/managed interop layer will allocate a byte array to marshal a native buffer to the `StreamDescriptor.Read` delegate. By using `Span<byte>` and the corresponding overload of `Stream.Read`, we can read directly into the caller's buffer when `CManagedStream::Read` calls this delegate.

Sample improvement by running a benchmark to create a `BitmapImage` with a `MemoryStream` `StreamSource` containing a 79KB PNG:

```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.1706 (21H2)
Intel Core i7-10875H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.100-preview.4.22252.9
  [Host]     : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT
  Job-FFRIND : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT

Runtime=.NET 6.0  Toolchain=net6.0-windows
```

|     Method |     Mean |     Error |    StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|----------- |---------:|----------:|----------:|---------:|---------:|---------:|----------:|
| Before | 3.623 ms | 0.0702 ms | 0.0656 ms | 984.3750 | 984.3750 | 984.3750 |     82 KB |
| After | 3.619 ms | 0.1386 ms | 0.4088 ms | 960.9375 | 960.9375 | 960.9375 |      3 KB |

## Customer Impact

Unnecessary allocation and memory copy between native/managed buffers.

## Regression

No

## Testing

Tested in real-world application via incorporation into [Faithlife.Wpf](https://www.nuget.org/packages/Faithlife.Wpf/).

## Risk

Minimal.

###### Microsoft Reviewers: [Open in CodeFlow](https://portal.fabricbot.ms/api/codeflow?pullrequest=https://github.com/dotnet/wpf/pull/6632)