@lindexi 2022-05-31
Although only 4 milliseconds have been reduced, it is important to reduce 79KB of memory allocated.