@bgrainger 2022-05-31
Actually only 4 microseconds faster, which is well within the margin of error. 

I'm not claiming that this PR is any faster; the primary goal is to eliminate managed allocations (which were previously equal in size to the input size of the image) and GCs.