@miloush 2022-06-04
No regression when the array is bigger than int? Doesn't AsSpan throw with negative numbers?