@bgrainger 2022-06-04
There's no regression because the old code cast the length to an `int`, too: https://github.com/dotnet/wpf/pull/6632/files#diff-43d84464891cb611646041f77d33aaf1c13800873ec0b7ff899f891985500d60L304