@bgrainger 2023-10-03
@dipeshmsft Could this PR be considered for the 8.x release series? It feels like a small change that has measurable allocation/GC benefits.